feat: complete first basic code
